requirements.txt
pyopengl
glfw
pillow
pyrr